httpd: upgrade.c: use mkstemp instead of dangerous mktemp; check for available memory first; correct argument in waitpid(); fix a few other issues

+++ release/src/router/shared/shutils.c
@@ -212,7 +212,7 @@ int safe_fread(void *ptr, size_t size, size_t nmemb, FILE *stream)
  * @param	stream	file stream
  * @return	number of items successfully written
  */
-int safe_fwrite(const void *ptr, size_t size, size_t nmemb, FILE *stream)
+size_t safe_fwrite(const void *ptr, size_t size, size_t nmemb, FILE *stream)
 {
 	size_t ret = 0;
 	do {
